+++ src/logic/Program.java
@@ -11,6 +11,10 @@ import drawing.JavaFXDrawingInformation;
 import geom.AbsolutePoint;
 import javafx.application.*;
 import javafx.stage.Stage;
+import logic.manager.DrawingManager;
+import logic.manager.KeyBoardManager;
+import logic.manager.MouseManager;
+import logic.manager.UpdateManager;
 import javafx.scene.*;
 
 import javafx.scene.paint.Color;
