+++ src/logic/manager/KeyBoardManager.java
@@ -1,4 +1,4 @@
-package logic;
+package logic.manager;
 
 import interfaces.logical.UpdatingObject;
