+++ src/space/shuttle/ArmedSpaceShuttle.java
@@ -3,8 +3,8 @@ package space.shuttle;
 import interfaces.geom.Point;
 import interfaces.logical.CollidingObject;
 import interfaces.logical.DestructibleObject;
-import logic.MouseManager;
-import logic.PlayerManager;
+import logic.manager.MouseManager;
+import logic.manager.PlayerManager;
 import space.core.SpaceObject;
 import space.shuttle.missiles.Laserbeam;
 import space.shuttle.missiles.Missile;
